refactor(runtime-v2): sink risk-calculator pure logic to core
- Move estimateLineChanges + FileModification from openclaw-plugin/src/core/risk-calculator.ts to principles-core/src/runtime-v2/risk/

- Remove dead code: assessRiskLevel, getTargetFileLineCount, calculatePercentageThreshold (Stage 2 bypass - zero production usage)

- Update gate.ts import to @principles/core/runtime-v2

- Update architecture-regression.test.ts (93 to 92) and core-anti-growth.test.ts

- Add risk/__tests__ to vitest.config.ts include patterns

ADR-0014 MVP-First compliant: pure refactor, no behavior change.
