+++ packages/principles-core/src/runtime-v2/__tests__/architecture-regression.test.ts
@@ -320,7 +320,6 @@ const KNOWN_PLUGIN_CORE_FILES = new Set([
   // with crypto.createHash + PainDictionary. Kept here because it owns the
   // crypto + dictionary I/O boundary.
   'detection-funnel.ts',
-  'risk-calculator.ts',
   'migration.ts',
   'file-store.ts',
   'pd-task-store.ts',
@@ -398,7 +397,7 @@ describe('PRI-212 plugin core anti-growth guard', () => {
     }
   });
 
-  it('known baseline count is self-consistent (94 files)', async () => {
+  it('known baseline count is self-consistent (93 files)', async () => {
     // Sanity check: if the baseline grows, update this number.
     // Prevents accidental baseline bloat from going unnoticed.
     // See docs/reviews/plugin-core-inventory-2026-05.md §7
@@ -420,7 +419,10 @@ describe('PRI-212 plugin core anti-growth guard', () => {
     // zero production references after refactor to @principles/core.
     // PRI-482: Added rule-context-assembler.ts (93 → 94) — plugin I/O boundary
     // converting TrajectoryDatabase rows to validated RuleHistoryWindow.
-    expect(KNOWN_PLUGIN_CORE_FILES.size).toBe(94);
+    // Stage-3: Removed risk-calculator.ts (94 → 93) — pure logic sunk to
+    // core runtime-v2/risk/, dead code (assessRiskLevel/getTargetFileLineCount/
+    // calculatePercentageThreshold) removed.
+    expect(KNOWN_PLUGIN_CORE_FILES.size).toBe(93);
   });
 });
