+++ packages/principles-core/src/runtime-v2/index.ts
@@ -1747,3 +1747,7 @@ export type {
 
 // PRI-470: IntentDecisionRecord durable SQLite store (SPEC §21.7).
 export { SqliteIntentDecisionStore } from './store/intent/index.js';
+
+// Risk calculator — pure line-change estimation migrated from plugin (Stage 3)
+export type { FileModification } from './risk/index.js';
+export { estimateLineChanges } from './risk/index.js';
