+++ packages/principles-core/src/runtime-v2/risk/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Risk — pure line-change estimation logic migrated from plugin.
+ * No I/O, no fs, no network.
+ */
+
+export type { FileModification } from './risk-calculator-policy.js';
+export { estimateLineChanges } from './risk-calculator-policy.js';
