+++ packages/principles-core/vitest.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     environment: 'node',
-    include: ['tests/**/*.test.ts', 'src/prompt-builder/__tests__/**/*.test.ts', 'src/quality-scorecard/__tests__/**/*.test.ts', 'src/runtime-v2/store/**/*.test.ts', 'src/runtime-v2/runner/**/*.test.ts', 'src/runtime-v2/utils/**/*.test.ts', 'src/runtime-v2/adapter/**/*.test.ts', 'src/runtime-v2/diagnostician/**/*.test.ts', 'src/runtime-v2/__tests__/**/*.test.ts', 'src/runtime-v2/gfi/__tests__/**/*.test.ts', 'src/runtime-v2/idle-trigger/__tests__/**/*.test.ts', 'src/runtime-v2/activation/__tests__/**/*.test.ts', 'src/runtime-v2/activation/writers/__tests__/**/*.test.ts', 'src/runtime-v2/internalization/__tests__/**/*.test.ts', 'src/runtime-v2/feature-flags/__tests__/**/*.test.ts', 'src/runtime-v2/observer/__tests__/**/*.test.ts', 'src/runtime-v2/config/__tests__/**/*.test.ts', 'src/runtime-v2/evidence-triage/__tests__/**/*.test.ts', 'src/runtime-v2/pain-gate/__tests__/**/*.test.ts', 'src/runtime-v2/detection/__tests__/**/*.test.ts', 'src/runtime-v2/core-principles/__tests__/**/*.test.ts', 'src/runtime-v2/tools/__tests__/**/*.test.ts', 'src/runtime-v2/feedback/__tests__/**/*.test.ts', 'src/runtime-v2/intent/__tests__/**/*.test.ts'],
+    include: ['tests/**/*.test.ts', 'src/prompt-builder/__tests__/**/*.test.ts', 'src/quality-scorecard/__tests__/**/*.test.ts', 'src/runtime-v2/store/**/*.test.ts', 'src/runtime-v2/runner/**/*.test.ts', 'src/runtime-v2/utils/**/*.test.ts', 'src/runtime-v2/adapter/**/*.test.ts', 'src/runtime-v2/diagnostician/**/*.test.ts', 'src/runtime-v2/__tests__/**/*.test.ts', 'src/runtime-v2/gfi/__tests__/**/*.test.ts', 'src/runtime-v2/idle-trigger/__tests__/**/*.test.ts', 'src/runtime-v2/activation/__tests__/**/*.test.ts', 'src/runtime-v2/activation/writers/__tests__/**/*.test.ts', 'src/runtime-v2/internalization/__tests__/**/*.test.ts', 'src/runtime-v2/feature-flags/__tests__/**/*.test.ts', 'src/runtime-v2/observer/__tests__/**/*.test.ts', 'src/runtime-v2/config/__tests__/**/*.test.ts', 'src/runtime-v2/evidence-triage/__tests__/**/*.test.ts', 'src/runtime-v2/pain-gate/__tests__/**/*.test.ts', 'src/runtime-v2/detection/__tests__/**/*.test.ts', 'src/runtime-v2/core-principles/__tests__/**/*.test.ts', 'src/runtime-v2/tools/__tests__/**/*.test.ts', 'src/runtime-v2/feedback/__tests__/**/*.test.ts', 'src/runtime-v2/intent/__tests__/**/*.test.ts', 'src/runtime-v2/risk/__tests__/**/*.test.ts'],
     coverage: {
       provider: 'v8',
       reporter: ['text', 'json', 'html'],
